requirements.txt
seaborn
branca
pybase64
pillow
databricks
databricks-sql-connector
streamlit
matplotlib
pandas
numpy
streamlit_option_menu
plotly
folium
streamlit_folium
geopandas
pyodbc
datetime
prophet
branca